update wasmtime dep and enable `concurrent_{imports|exports}`
Previously, the `wasm_runner` code had disabled the `concurrent_imports` and
`concurrent_exports` binding generation options, which prevented any kind of
guest-level concurrency.  Also, there seems to be a bug in the
`wasip3-prototyping` fork of Wasmtime such that calling an async-lifted export
using `TypedFunc::call_async` instead of `TypedFunc::call_concurrent` leads to a
"wasm trap: async-lifted export failed to produce a result" error.  That's not
super surprising since we haven't bothered to test that scenario yet, but we'll
need to address it eventually.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>